+++ subsys/bluetooth/mesh/app_keys.h
@@ -7,8 +7,42 @@
 #ifndef ZEPHYR_SUBSYS_BLUETOOTH_MESH_APP_KEYS_H_
 #define ZEPHYR_SUBSYS_BLUETOOTH_MESH_APP_KEYS_H_
 
+#include <zephyr/sys/iterable_sections.h>
 #include <zephyr/bluetooth/mesh.h>
 #include "subnet.h"
+
+/** Mesh Application Key. */
+struct app_key {
+	/** NetKey Index the AppKey is bound to. */
+	uint16_t net_idx;
+	/** Application Key Index. */
+	uint16_t app_idx;
+	/** True if the AppKey has a second key for KR. */
+	bool updated;
+	struct bt_mesh_app_cred {
+		/** Application ID. */
+		uint8_t id;
+		/** Application Key value. */
+		struct bt_mesh_key val;
+	} keys[2];
+};
+
+/** Appkey callback. Instantiate with @ref BT_MESH_APP_KEY_CB_DEFINE */
+struct bt_mesh_app_key_cb {
+	void (*evt_handler)(struct app_key *app, enum bt_mesh_key_evt evt);
+};
+
+/**
+ *  @brief Register an AppKey event callback.
+ *
+ *  @param _handler Handler function, see @ref bt_mesh_app_key_cb::evt_handler.
+ */
+#define BT_MESH_APP_KEY_CB_DEFINE(_handler)                                    \
+	static const STRUCT_SECTION_ITERABLE(bt_mesh_app_key_cb,               \
+					     _CONCAT(bt_mesh_app_key_cb_,      \
+						       _handler)) = {          \
+		.evt_handler = (_handler),                                     \
+	}
 
 /** @brief Reset the app keys module. */
 void bt_mesh_app_keys_reset(void);
