+++ subsys/bluetooth/mesh/cfg_srv.c
@@ -22,6 +22,7 @@
 
 #include "mesh.h"
 #include "net.h"
+#include "app_keys.h"
 #include "rpl.h"
 #include "lpn.h"
 #include "transport.h"
@@ -343,11 +344,10 @@ static void mod_app_key_del(const struct bt_mesh_model *mod,
 	mod_unbind(mod, *app_idx, true);
 }
 
-static void app_key_evt(uint16_t app_idx, uint16_t net_idx,
-			enum bt_mesh_key_evt evt)
+static void app_key_evt(struct app_key *app, enum bt_mesh_key_evt evt)
 {
 	if (evt == BT_MESH_KEY_DELETED) {
-		bt_mesh_model_foreach(&mod_app_key_del, &app_idx);
+		bt_mesh_model_foreach(&mod_app_key_del, &app->app_idx);
 	}
 }
